+++ signin.html
@@ -14,7 +14,7 @@
       type="text/css"
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css"
     />
-    <script defer type="module" src="./src/sign.js"></script>
+    <script defer type="module" src="./src/signIn.js"></script>
   </head>
   <body>
     <div class="main">
@@ -25,7 +25,7 @@
           </a>
           <div class="subtitle">
             <span>회원이 아니신가요?</span>
-            <a href="./signup.html">회원 가입하기</a>
+            <a href="./signUp.html">회원 가입하기</a>
           </div>
         </div>
         <form>
@@ -44,7 +44,6 @@
                   id="pw-eyeIcon"
                   src="./public/images/sign/eye-off.png"
                   alt="비밀번호 숨기기 아이콘"
-                  onclick="togglePwVisibility()"
                 />
               </div>
             </div>
