+++ signup.html
@@ -25,7 +25,7 @@
           </a>
           <div class="subtitle">
             <span>이미 회원이신가요?</span>
-            <a href="./signin.html">로그인 하기</a>
+            <a href="./signIn.html">로그인 하기</a>
           </div>
         </div>
         <form>
@@ -44,7 +44,6 @@
                   id="pw-eyeIcon"
                   src="./public/images/sign/eye-off.png"
                   alt="비밀번호 숨기기 아이콘"
-                  onclick="togglePwVisibility()"
                 />
               </div>
             </div>
@@ -60,7 +59,6 @@
                   id="pwConfirm-eyeIcon"
                   src="./public/images/sign/eye-off.png"
                   alt="비밀번호 숨기기 아이콘"
-                  onclick="togglePwConfirmVisibility()"
                 />
               </div>
             </div>
